+++ CHANGELOG.md
@@ -4,7 +4,7 @@ To see version information, `xsvutils --version`.
 
 To see compatibility information, `xsvutils help compatibility`.
 
-## version 0.3.3 (Unreleased)
+## version 0.3.3 (2018/05/27)
 - add `rmnoname` subcommand
 - add `insunixtime` subcommand
 - add `--last` option of `cols` subcommand